+++ packages/ludiek/src/engine/LudiekFeature.ts
@@ -1,5 +1,5 @@
 import { LudiekElement } from '@ludiek/engine/LudiekElement';
-import { LudiekDependencies } from '@ludiek/engine/LudiekEngineConcept';
+import { LudiekDependencies } from './LudiekEngineContribution';
 
 export abstract class LudiekFeature<
   Dependencies extends LudiekDependencies = object,
