+++ packages/ludiek/src/engine/request/LudiekRequest.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { LudiekDependencies, LudiekEngineConcept } from '@ludiek/engine/LudiekEngineConcept';
+import { LudiekDependencies, LudiekEngineContribution } from '../LudiekEngineContribution';
 import { IsNonEmpty } from '@ludiek/util/types';
 
 /**
@@ -14,7 +14,7 @@ export type BaseRequest = z.infer<typeof BaseRequestSchema>;
 export abstract class LudiekController<
   Request extends BaseRequest = BaseRequest,
   Dependencies extends LudiekDependencies = object,
-> extends LudiekEngineConcept<Dependencies> {
+> extends LudiekEngineContribution<Dependencies> {
   declare readonly __request: Request;
 
   public abstract readonly schema: z.ZodObject<{
